make game a singleton + ensure a single game is launched, throw if attempting another one + consolidate graphics context to a single class + hide away Silk and Veldrid's enums with own enum
